# artifacts/handoff-w4-round-1/provider-artifacts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/organizations/{organizationId}": {
      "get": {
        "operationId": "OrganizationsController_get",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Organizations"
        ]
      },
      "patch": {
        "operationId": "OrganizationsController_update",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Organizations"
        ]
      }
    },
...
    "/v1/organizations/{organizationId}/customers": {
      "get": {
        "operationId": "MerchantOperationsController_customers",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipStatus",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rewardReady",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/customers/{customerId}": {
      "get": {
        "operationId": "MerchantOperationsController_customer",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/memberships/{membershipId}": {
      "get": {
        "operationId": "MerchantOperationsController_membership",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/memberships/{membershipId}/ledger": {
      "get": {
        "operationId": "MerchantOperationsController_ledger",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/memberships/{membershipId}/rewards": {
      "get": {
        "operationId": "MerchantOperationsController_rewards",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/memberships/{membershipId}/suspend": {
      "post": {
        "operationId": "MerchantOperationsController_suspend",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/memberships/{membershipId}/restore": {
      "post": {
        "operationId": "MerchantOperationsController_restore",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/memberships/{membershipId}/revoke": {
      "post": {
        "operationId": "MerchantOperationsController_revoke",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/memberships/{membershipId}/manual-adjustment": {
      "post": {
        "operationId": "MerchantOperationsController_manualAdjustment",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/memberships/{membershipId}/verify-projection": {
      "post": {
        "operationId": "MerchantOperationsController_verifyProjection",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/memberships/{membershipId}/rebuild-projection": {
      "post": {
        "operationId": "MerchantOperationsController_rebuildProjection",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/operation-approvals": {
      "post": {
        "operationId": "MerchantOperationsController_createApproval",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      },
      "get": {
        "operationId": "MerchantOperationsController_approvals",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/operation-approvals/{approvalId}/approve": {
      "post": {
        "operationId": "MerchantOperationsController_approve",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "approvalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/operation-approvals/{approvalId}/reject": {
      "post": {
        "operationId": "MerchantOperationsController_reject",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "approvalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/risk-signals": {
      "get": {
        "operationId": "MerchantOperationsController_riskSignals",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "severity",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/risk-signals/{signalId}": {
      "get": {
        "operationId": "MerchantOperationsController_riskSignal",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "signalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/risk-signals/{signalId}/acknowledge": {
      "post": {
        "operationId": "MerchantOperationsController_acknowledgeRisk",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "signalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
    "/v1/organizations/{organizationId}/risk-signals/{signalId}/resolve": {
      "post": {
        "operationId": "MerchantOperationsController_resolveRisk",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "signalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantOperations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Waflo Platform API",
    "description": "Authoritative Waflo W4 API for merchant administration and signed Staff devices.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "waflo_session"
      }
    },
...
  }
}